+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#191][GL-191] - Extra apples from the apple-drops option are only dropped when leaves decay and not when a player breaks leaves
 - [#192][GL-192] - All types of decaying leaves drop oak saplings when the apple-drops.all-trees option is enabled
 - [#193][GL-193] - The apple-drops.all-trees option can only be disabled on Minecraft 1.13+
+- [#187][GL-187] - Only the first tool is enchanted when shift-clicking recipe result in the Hastey Boys scenario
 - [#184][GL-184] - Various bugs and compatibility issues between scenarios with custom block drops
 - [#180][GL-180] - Player death message is not client-translatable on Paper servers for Minecraft 1.16.5+
 - [#177][GL-177] - Iteminfo command fails when using it on a potion that has no base effect on Minecraft 1.20.5+
@@ -22,6 +23,7 @@
 [GL-192]: https://gitlab.com/uhccore/uhccore/-/issues/192
 [GL-191]: https://gitlab.com/uhccore/uhccore/-/issues/191
 [GL-190]: https://gitlab.com/uhccore/uhccore/-/issues/190
+[GL-187]: https://gitlab.com/uhccore/uhccore/-/issues/187
 [GL-184]: https://gitlab.com/uhccore/uhccore/-/issues/184
 [GL-182]: https://gitlab.com/uhccore/uhccore/-/issues/182
 [GL-180]: https://gitlab.com/uhccore/uhccore/-/issues/180